cleaned up shader names and code style

+++ public/shaders/texture_D/fragment.glsl
@@ -1,4 +1,4 @@
-varying highp vec2 vTextureCoord;
+varying mediump vec2 vTextureCoord;
 
 uniform sampler2D uSamplerDiffuse;
